# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local Development"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/Admin/Users/LoadData": {
            "post": {
                "tags": [
                    "Admins"
                ],
                "summary": "Get all users",
                "description": "Returns a paginated list of admins with search and sorting capabilities. Includes no, name, username, role, email, supplier_id, and created_at.",
                "operationId": "fe2b82bd0062a0ed885cb30e2ca12cf1",
                "requestBody": {
                    "description": "Optional parameters for filtering and pagination",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "search": {
                                        "description": "Global search across name, email, and username",
                                        "type": "string",
                                        "example": "john"
                                    },
                                    "sortColumn": {
                                        "description": "Column to sort by (fullName, username, email, role, created_at, id)",
                                        "type": "string",
                                        "example": "created_at"
                                    },
                                    "sortColumnDir": {
                                        "description": "Sort direction",
                                        "type": "string",
                                        "enum": [
                                            "asc",
                                            "desc"
                                        ],
                                        "example": "desc"
                                    },
                                    "pageNumber": {
                                        "description": "Page number",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "pageSize": {
                                        "description": "Items per page",
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "role": {
                                        "description": "Filter by role (slug or name, case-insensitive)",
                                        "type": "string",
                                        "example": "PEMASOK"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "no": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "id": {
                                                        "type": "string",
                                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                                    },
                                                    "firstName": {
                                                        "type": "string",
                                                        "example": "John"
                                                    },
                                                    "lastName": {
                                                        "type": "string",
                                                        "example": "Doe"
                                                    },
                                                    "fullName": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "username": {
                                                        "type": "string",
                                                        "example": "johndoe"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "john@example.com"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "Super Admin"
                                                    },
                                                    "roles": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "Super Admin"
                                                                },
                                                                "slug": {
                                                                    "type": "string",
                                                                    "example": "super_admin"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "supplierId": {
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "active"
                                                    },
                                                    "isSuperAdmin": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "createdAt": {
                                                        "type": "string",
                                                        "example": "2025-01-15 10:30:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "pageSize": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "totalPages": {
                                                    "type": "integer",
                                                    "example": 4
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}